pyproject.toml
[project]
name = "probefs"
version = "0.1.0"
description = "A fast, beautiful, keyboard-driven TUI file browser"
readme = "README.md"
authors = [
    { name = "Dale Wright", email = "dalwrigh@cisco.com" }
]
requires-python = ">=3.10"
dependencies = [
    "textual",
    "fsspec",
    "ruamel-yaml>=0.19.1",
    "send2trash>=2.1.0",
]

[project.scripts]
probefs = "probefs.app:main"

[project.optional-dependencies]
sftp = ["paramiko>=3.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-textual-snapshot>=1.0.0",
]
